requirements.txt
Django==1.8.13
django-braces==1.9.0
django-ckeditor==5.0.3
django-crispy-forms==1.6.0
django-endless-pagination==2.0
mysqlclient==1.3.7
Pillow==3.3.0
pkg-resources==0.0.0
psycopg2==2.6.1
